Store 2 unrelated GetSize() results in appropriate data types.
Avoid storing results in 32-bit ints, when the results can potentially
be bigger.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/96611
Commit-Queue: Lei Zhang <thestig@chromium.org>

+++ core/fpdfapi/parser/cpdf_crypto_handler.cpp
@@ -144,7 +144,7 @@ bool CPDF_CryptoHandler::DecryptStream(void* context,
     return true;
   }
   if (m_Cipher == Cipher::kRC4) {
-    int old_size = dest_buf.GetSize();
+    size_t old_size = dest_buf.GetSize();
     dest_buf.AppendBlock(source.data(), source.size());
     CRYPT_ArcFourCrypt(static_cast<CRYPT_rc4_context*>(context),
                        dest_buf.GetSpan().subspan(old_size, source.size()));
